housekeeping: Remove not used files: - markdown2html is not used (we are using "Sundown" tool) - newstroke_font_without_CJK.cpp is removed. Due to code optimization the full font works fine without memory issues -fix also a few very minor compil warnings

+++ eeschema/sim/sim_plot_panel.cpp
@@ -648,7 +648,7 @@ wxColour SIM_PLOT_PANEL::generateColor()
 
     const unsigned int colorCount = sizeof(colors) / sizeof(unsigned long);
 
-    for( int i = 0; i < colorCount - 1; i++ )
+    for( int i = 0; i < (int)colorCount - 1; i++ )
     {
         const wxColour color = wxColour( colors[i] );
         bool hasColor = false;
